fix hw 02
