+++ public/server.log
@@ -9,3 +9,126 @@ PUT /api/contacts/sopPIPEzhodBHkRiXfTnB 13.04.2023_04:48:46
 DELETE /api/contacts/sopPIPEzhodBHkRiXfTnB 13.04.2023_04:56:05
 PUT /api/contacts/sopPIPEzhodBHkRiXfTnB 13.04.2023_04:56:15
 POST /api/contacts 13.04.2023_04:56:47
+POST /api/contacts 14.04.2023_06:37:20
+POST /api/contacts 14.04.2023_06:37:28
+POST /api/contacts 14.04.2023_06:39:01
+POST /api/contacts 14.04.2023_06:41:30
+POST /api/contacts 14.04.2023_06:41:31
+POST /api/contacts 14.04.2023_06:41:36
+POST /api/contacts 14.04.2023_06:41:53
+POST /api/contacts 14.04.2023_06:41:59
+POST /api/contacts 14.04.2023_06:44:51
+POST /api/contacts 14.04.2023_06:44:52
+GET /api/contacts 14.04.2023_06:45:08
+POST /api/contacts 14.04.2023_06:45:28
+POST /api/contacts 14.04.2023_06:47:06
+POST /api/contacts 14.04.2023_06:47:07
+POST /api/contacts 14.04.2023_06:48:25
+POST /api/contacts 14.04.2023_06:48:27
+GET /api/contacts 14.04.2023_06:49:41
+POST /api/contacts 14.04.2023_06:49:51
+PUT /api/contacts 14.04.2023_06:49:57
+PUT /api/contacts 14.04.2023_06:50:03
+PUT /api/contacts 14.04.2023_06:50:20
+PUT /api/contacts 14.04.2023_06:50:38
+PUT /api/contacts 14.04.2023_06:50:39
+PUT /api/contacts 14.04.2023_06:51:08
+PUT /api/contacts 14.04.2023_06:51:10
+POST /api/contacts 14.04.2023_06:51:14
+POST /api/contacts 14.04.2023_06:51:16
+POST /api/contacts 14.04.2023_06:51:17
+POST /api/contacts 14.04.2023_06:51:18
+POST /api/contacts 14.04.2023_06:51:56
+POST /api/contacts 14.04.2023_07:00:27
+POST /api/contacts 14.04.2023_07:00:28
+POST /api/contacts 14.04.2023_07:00:59
+POST /api/contacts 14.04.2023_07:01:14
+POST /api/contacts 14.04.2023_07:01:23
+POST /api/contacts 14.04.2023_07:05:30
+POST /api/contacts 14.04.2023_07:05:31
+POST /api/contacts 14.04.2023_07:05:32
+POST /api/contacts 14.04.2023_07:06:34
+POST /api/contacts 14.04.2023_07:06:36
+POST /api/contacts 14.04.2023_07:07:44
+POST /api/contacts 14.04.2023_07:07:45
+POST /api/contacts 14.04.2023_07:07:52
+POST /api/contacts 14.04.2023_07:09:32
+POST /api/contacts 14.04.2023_07:09:52
+POST /api/contacts 14.04.2023_07:09:54
+POST /api/contacts 14.04.2023_07:09:55
+POST /api/contacts 14.04.2023_07:10:04
+POST /api/contacts 14.04.2023_07:15:17
+POST /api/contacts 14.04.2023_07:15:47
+POST /api/contacts 14.04.2023_07:15:55
+POST /api/contacts 14.04.2023_07:16:51
+POST /api/contacts 14.04.2023_07:17:44
+POST /api/contacts 14.04.2023_07:17:53
+POST /api/contacts 14.04.2023_07:18:49
+POST /api/contacts 14.04.2023_07:18:55
+POST /api/contacts 14.04.2023_07:20:30
+POST /api/contacts 14.04.2023_07:20:31
+POST /api/contacts 14.04.2023_07:20:32
+POST /api/contacts 14.04.2023_07:20:33
+POST /api/contacts 14.04.2023_07:20:34
+POST /api/contacts 14.04.2023_07:21:07
+POST /api/contacts 14.04.2023_07:21:08
+POST /api/contacts 14.04.2023_07:21:23
+POST /api/contacts 14.04.2023_07:21:33
+POST /api/contacts 14.04.2023_07:21:37
+POST /api/contacts 14.04.2023_07:21:47
+POST /api/contacts 14.04.2023_07:21:56
+POST /api/contacts 14.04.2023_07:22:10
+POST /api/contacts 14.04.2023_07:22:15
+POST /api/contacts 14.04.2023_07:23:59
+POST /api/contacts 14.04.2023_07:24:00
+POST /api/contacts 14.04.2023_07:26:10
+POST /api/contacts 14.04.2023_07:26:11
+POST /api/contacts 14.04.2023_07:26:28
+POST /api/contacts 14.04.2023_07:26:44
+POST /api/contacts 14.04.2023_07:26:57
+POST /api/contacts 14.04.2023_07:27:16
+POST /api/contacts 14.04.2023_07:27:17
+POST /api/contacts 14.04.2023_07:27:45
+POST /api/contacts 14.04.2023_07:28:12
+POST /api/contacts 14.04.2023_07:29:09
+POST /api/contacts 14.04.2023_07:30:22
+POST /api/contacts 14.04.2023_07:30:39
+POST /api/contacts 14.04.2023_07:33:24
+POST /api/contacts 14.04.2023_07:33:38
+POST /api/contacts 14.04.2023_07:34:03
+POST /api/contacts 14.04.2023_07:34:44
+POST /api/contacts 14.04.2023_07:34:50
+POST /api/contacts 14.04.2023_07:34:52
+POST /api/contacts 14.04.2023_07:35:03
+POST /api/contacts 14.04.2023_07:35:51
+POST /api/contacts 14.04.2023_07:37:41
+POST /api/contacts 14.04.2023_07:37:48
+POST /api/contacts 14.04.2023_07:37:57
+POST /api/contacts 14.04.2023_07:38:11
+PUT /api/contacts/1 14.04.2023_07:42:35
+PUT /api/contacts/1 14.04.2023_07:44:45
+PUT /api/contacts/1 14.04.2023_07:51:19
+PUT /api/contacts/1 14.04.2023_07:52:00
+PUT /api/contacts/1 14.04.2023_07:52:01
+PUT /api/contacts/4vdbf3nolggql26r 14.04.2023_07:52:19
+PUT /api/contacts/4vdbf3nolggql26f 14.04.2023_07:52:26
+PUT /api/contacts/4vdbf3nolggql26f 14.04.2023_07:56:50
+PUT /api/contacts/4vdbf3nolggql26f 14.04.2023_07:56:54
+POST /api/contacts/4vdbf3nolggql26f 14.04.2023_07:56:59
+POST /api/contacts/ 14.04.2023_07:57:15
+POST /api/contacts/ 14.04.2023_07:57:21
+GET /api/contacts/ 14.04.2023_08:19:45
+GET /api/contacts/ 14.04.2023_09:09:30
+POST /api/contacts/ 14.04.2023_09:09:43
+POST /api/contacts/ 14.04.2023_09:11:13
+POST /api/contacts/ 14.04.2023_09:11:19
+POST /api/contacts/ 14.04.2023_09:11:54
+POST /api/contacts/ 14.04.2023_09:12:01
+POST /api/contacts/ 14.04.2023_09:12:06
+POST /api/contacts/ 14.04.2023_09:12:49
+POST /api/contacts/ 14.04.2023_09:12:56
+POST /api/contacts/ 14.04.2023_09:13:09
+PUT /api/contacts/4vdbfp14lggvapfb 14.04.2023_09:13:31
+PUT /api/contacts/4vdbfp14lggvapfb7 14.04.2023_09:13:41
+PUT /api/contacts/4vdbfp14lggvapfb 14.04.2023_09:13:46
+DELETE /api/contacts/4vdbfp14lggvapfb 14.04.2023_09:13:52
