+++ schemas/schemas.js
@@ -1,0 +1,11 @@
+const Joi = require("joi");
+const contacts = require("../models/contacts");
+const contactSchema = Joi.object({
+  name: Joi.string().alphanum().min(3).max(30).required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
+    .required(),
+  phone: Joi.string().required(),
+});
+
+module.exports = contactSchema;
